Major update moving away from Message class to functions accessible by include
