+++ include/gestalt/gestalt.h
@@ -1,0 +1,10 @@
+/*
+ * gestalt.h
+ *
+ * This file contains the definitions of ALL Gestalts security functions.
+ * 
+ * Author: Hunter L, Richardson
+ * Date: 2024-03-07
+ */
+
+//#include <gestalt/aes.h>
