+++ src/modes/modes.h
@@ -11,19 +11,20 @@
 #pragma once
 
 #include "../../tools/utils.h"
-#include "../aes/aes.h"
+#include "../aes/aesCore.h"
 
 #include <string>
+#include <cstring>
 
 template<typename BlockCipher>
 using function = void (BlockCipher::*)(unsigned char*);
 
 template<typename BlockCipher>
-void encrypt_ecb(std::string& msg, std::string key, function<BlockCipher> encryptBlock);
+std::string encrypt_ecb(std::string& msg, std::string key, function<BlockCipher> encryptBlock);
 template<typename BlockCipher>
-void decrypt_ecb(std::string& msg, std::string key, function<BlockCipher> decryptBlock);
+std::string decrypt_ecb(std::string& msg, std::string key, function<BlockCipher> decryptBlock);
 
 template<typename BlockCipher>
-void encrypt_cbc(std::string& msg, std::string key, std::string iv, function<BlockCipher> encryptBlock);
+std::string encrypt_cbc(std::string& msg, std::string key, std::string iv, function<BlockCipher> encryptBlock);
 template<typename BlockCipher>
-void decrypt_cbc(std::string& msg, std::string key, std::string iv, function<BlockCipher> decryptBlock);
+std::string decrypt_cbc(std::string& msg, std::string key, std::string iv, function<BlockCipher> decryptBlock);
