+++ tests/CMakeLists.txt
@@ -1,14 +1,15 @@
-cmake_minimum_required(VERSION 3.27.2)
+cmake_minimum_required(VERSION 3.16.3)
 
 include(GoogleTest)
 
 set (This tests)
 
 set (Sources
-    tests.cpp
+    aesTests.cpp
 )
 
 add_executable (${This} ${Sources})
+
 target_link_libraries (${This} PRIVATE
     GTest::gtest_main
     Gestalt
@@ -20,4 +21,4 @@ add_test (
 )
 
 include(GoogleTest)
-gtest_discover_tests(tests)
+gtest_discover_tests(${This})
