+++ tests/desTests.cpp
@@ -1,0 +1,7 @@
+
+#include "gtest/gtest.h"
+
+TEST(sample_test_case, sample_test)
+{
+    EXPECT_EQ(1, 1);
+}
